+++ website/pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
-import initAuth from '../logic/initAuth'
+import initAuth from '../utils/initAuth'
 
 initAuth()
